Add files via upload
A PHP program to calculate the mathematical function 3x + 1
First the program works with a single number and print all the values, find max value (Highest number) and count of iterations (Total stopping time).
then another function is created for calculating function values ​​(max value and count of iterations) of numbers range (eg, 10 to 10^6). 
The HTML Forms was created and prompted for range values ​​(start and finish). As results store in array: numbers, maximum (highest) values ​​and the number of iterations of them.
it finds the numbers with max and min iterations. Print both (2) number, count of iteration and highest values.